finish modifications to aspect ratio from Martin BC. the "boundary" parameter now accepts the following options: - "linked": twist-and-shift BC, with generalization for non-axisymmetric geometry (Martin et al 2018) - "forced periodic" (or simply "periodic"): use periodic BCs with no cutting of flux tube - "exact periodic": cut flux tube at a location where gds21 = 0, and then use periodic BCs (currently for VMEC geometry only) - "continuous drifts": cut flux tube at a location where gbdrift0 = 0, and then use (generalized) twist-and-shift BC (currently for VMEC geometry only) also, npol can now be a float. and which_crossing now defaults to -1, which takes the outermost cut (given npol). finally, I think jtwist should be signed, because kx_shift can be positive or negative, depending on shear. and so this is now how it is implemented.

+++ geometry_modules/vmec/include/geometric_coefficients.h
@@ -212,7 +212,7 @@ class Geometric_coefficients {
   double alpha = 0.0;
   int nzgrid = 16;
   int nzgrid_cut = 16;
-  int npol = 1;
+  double npol = 1;
   int sign_psi;
   double desired_normalized_toroidal_flux = 0.25;
   int vmec_surface_option = 2;
